+++ src/Shiny.Core/Power/Platforms/Android/PowerManagerImpl.cs
@@ -36,9 +36,18 @@ namespace Shiny.Power
         }
 
 
+        public bool IsEnergySavingEnabled
+        {
+            get
+            {
+                
+                using (var pm = (PowerManager)this.context.AppContext.GetSystemService(Context.PowerService))
+                    return pm.IsDeviceIdleMode && !pm.IsIgnoringBatteryOptimizations(this.context.Package.PackageName);
+            }
+        }
         public PowerState Status => this.context.GetIntentValue(Intent.ActionBatteryChanged, ToState);
         public int BatteryLevel => this.context.GetIntentValue(Intent.ActionBatteryChanged, ToLevel);
-
+        
 
         static PowerState ToState(Intent intent)
         {