+++ src/Shiny.Core/Power/Platforms/iOS/PowerManagerImpl.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
+using Foundation;
 using UIKit;
 
 
@@ -11,9 +12,11 @@ namespace Shiny.Power
         CompositeDisposable? dispose;
 
 
+        public bool IsEnergySavingEnabled => NSProcessInfo.ProcessInfo.LowPowerModeEnabled;
+
+
         protected override void OnNpcHookChanged(bool hasSubscribers)
         {
-
             if (hasSubscribers)
             {
                 this.dispose = new CompositeDisposable(
@@ -53,8 +56,7 @@ namespace Shiny.Power
                 }
             }
         }
-
-
+       
 
         IObservable<int> WhenBatteryPercentageChanged() => Observable.Create<int>(ob =>
         {